+++ Source/Core/Runtime/Properties/BaseHighlightProperty.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -7,7 +7,7 @@ namespace VRBuilder.Core.Properties
     /// <summary>
     /// Abstract base property for highlight properties.
     /// </summary>
-    public abstract class BaseHighlightProperty : ProcessSceneObjectProperty, IHighlightProperty
+    public abstract class BaseHighlightProperty : ColorHighlightPropertyBase<HighlightPropertyEventArgs>, IHighlightProperty
     {
         [Header("Events")]
         [SerializeField]
@@ -19,12 +19,14 @@ namespace VRBuilder.Core.Properties
         /// <summary>
         /// Event data for events of <see cref="BaseHighlightProperty"/>.
         /// </summary>
-        public class HighlightEventArgs : EventArgs { }
+        public class HighlightEventArgs : EventArgs
+        {
+        }
 
         /// <summary>
         /// Is currently highlighted.
         /// </summary>
-        public bool IsHighlighted { get; protected set; }
+        public bool IsHighlighted => IsActive;
 
         /// <inheritdoc/>
         public UnityEvent<HighlightPropertyEventArgs> HighlightStarted => highlightStarted;
@@ -32,26 +34,50 @@ namespace VRBuilder.Core.Properties
         /// <inheritdoc/>
         public UnityEvent<HighlightPropertyEventArgs> HighlightEnded => highlightEnded;
 
-        /// <inheritdoc/>
-        public abstract void Highlight(Color highlightColor);
+        /// <inheritdoc />
+        protected override UnityEvent<HighlightPropertyEventArgs> StartedEvent => HighlightStarted;
+
+        /// <inheritdoc />
+        protected override UnityEvent<HighlightPropertyEventArgs> EndedEvent => HighlightEnded;
 
         /// <inheritdoc/>
-        public abstract void Unhighlight();
-
-        /// <summary>
-        /// Emits an event when the property is highlighted.
-        /// </summary>
-        public void EmitHighlightEvent(HighlightPropertyEventArgs args)
+        public virtual void Highlight(Color highlightColor)
         {
-            HighlightStarted?.Invoke(args);
+            Activate(highlightColor);
+        }
+
+        /// <inheritdoc/>
+        public virtual void Unhighlight()
+        {
+            Deactivate();
         }
 
         /// <summary>
-        /// Emits an event when the property is unhighlighted.
+        /// Applies the visual highlighted state.
         /// </summary>
-        public void EmitUnhighlightEvent(HighlightPropertyEventArgs args)
+        protected abstract bool TryHighlight(Color highlightColor);
+
+        /// <summary>
+        /// Applies the visual unhighlighted state.
+        /// </summary>
+        protected abstract bool TryUnhighlight();
+
+        /// <inheritdoc />
+        protected sealed override bool TryApplyVisualState(bool isActive, Color color)
         {
-            HighlightEnded?.Invoke(args);
+            return isActive ? TryHighlight(color) : TryUnhighlight();
+        }
+
+        /// <inheritdoc />
+        protected override Color? GetEndedEventColor(Color? color)
+        {
+            return null;
+        }
+
+        /// <inheritdoc />
+        protected override HighlightPropertyEventArgs CreateEventArgs(Color? color)
+        {
+            return new HighlightPropertyEventArgs(color);
         }
     }
 }